# LayerList performance improvements