--- a/config/webpack.production.config.js
+++ b/config/webpack.production.config.js
@@ -18,6 +18,9 @@ module.exports = {
     filename: '[name].[contenthash].js',
     chunkFilename: '[contenthash].js'
   },
+  optimization:{
+    minimize: false,
+  },
   resolve: {
     extensions: ['.js', '.jsx']
   },