@pathmapper 2019-10-13
Hmm, do you get consistent values for multiple test runs @orangemug ?

Asking because I run multiple test with production builds of `v1.6.0`, `v.1.6.1` and your `demo build` and can't say that one of these gives better results than the others.

Used the latest versions of Firefox and Chome (with Firefox the range of the results is wider).