@pathmapper 2019-10-14
Here are the results of 10 test runs for each production build. All on Chrome, all the same test setup.
Test: Changing the background color of osm-liberty using the color picker (click)

|Test run | ms (PR demo) | ms (v1.6.1) | ms (v1.6.0) |
| :------------- | -------------: | -------------: | -------------: |
| 1 |  100.1 | 115.8 | 103.0
| 2|  106.2 | 61.9| 101.4
| 3 |  83.7 | 97.4| 94.0
|  4|   100.4| 91.4| 94.1
|  5|  110.7 | 148.2| 77.3
|  6|   58.8| 92.9| 92.6
|  7|   63.9| 74.5| 81.4
|  8|   85.2| 53.8| 96.5
|  9|  79.3 | 85.7| 86.5
|  10|   78.0| 82.5| 90.4