@orangemug 2019-10-14
Ok so here's where I'm at with this PR. So my initial condition wasn't very good (sorry) background color will trigger a color transition in mapbox-gl-js. So depending on what color you pick you get different performance from mapbox-gl-js, which will in turn effect frame rate. Ideally we want to just test the react/UI bit, the stuff we have control over.  

So I've changed my condition for the performance metrics I'm now changing the 'max zoom' of the 'state' layer osm-liberty. In a way that it would cause a re-render of the mapbox-gl map. So for example with the current map zoom at 1, I would change the 'max zoom' of the 'state' layer between `6` and `7`

I've also install the react chrome extension, so I can do profiling in that instead in development. Here are my development results

Without `shouldComponentUpdate`

<img width="589" alt="Screen Shot 2019-10-14 at 10 09 42" src="https://user-images.githubusercontent.com/235915/66748402-0134b700-ee7f-11e9-86ba-4f272c678de7.png">
<img width="588" alt="Screen Shot 2019-10-14 at 10 09 21" src="https://user-images.githubusercontent.com/235915/66748403-0134b700-ee7f-11e9-8aee-4b8b6c5bd44e.png">
<img width="588" alt="Screen Shot 2019-10-14 at 10 08 53" src="https://user-images.githubusercontent.com/235915/66748404-0134b700-ee7f-11e9-8368-fbef750965ba.png">

With `shouldComponentUpdate`
    
<img width="588" alt="Screen Shot 2019-10-14 at 10 11 19" src="https://user-images.githubusercontent.com/235915/66748435-16a9e100-ee7f-11e9-87f4-7c7800ead5bc.png">
<img width="589" alt="Screen Shot 2019-10-14 at 10 11 02" src="https://user-images.githubusercontent.com/235915/66748436-16a9e100-ee7f-11e9-96a7-2a3baffcd981.png">
<img width="589" alt="Screen Shot 2019-10-14 at 10 10 26" src="https://user-images.githubusercontent.com/235915/66748437-16a9e100-ee7f-11e9-9075-c660ec98d190.png">

I then built the 2 different versions and tested with the regular chrome devtools performance tab. I tested 8x3 times to reduce the effect of the CPU being busy with background tasks.  

```
irb(main):004:0> (85.0+74+82+110+124+91+94+109)/8 # new code
=> 96.125
irb(main):001:0> (108.0+82+97+78+135+149+156+111)/8  #1.6.1
=> 114.5
irb(main):006:0> (99.0+134+108+110+115+100+83+158)/8 # new code
=> 113.375
irb(main):008:0> (141.0+136+130+95+144+130+116+97)/8 # 1.6.1
=> 123.625
irb(main):009:0> (81.0+71+95+74+75+86+78+88)/8  # new code
=> 81.0
irb(main):010:0> (137.0+135+103+104+116+110+125+97)/8  # 1.6.1
=> 115.875 
```

So if we average across the test runs for the new code and 1.6.1 I get the following results

```
irb(main):002:0> (96.125+113.375+81.0)/3  # new code
=> 96.83333333333333
irb(main):003:0> (114.5+123.625+115.875)/3  # 1.6.1
=> 118.0
```

Hardly amazing results but 20ms is better than nothing.

Next steps

 - I think I'll add a [react profiling-in-production](https://gist.github.com/bvaughn/25e6233aeb1b4f0cdb8d8366e54a3977#profiling-in-production) as a separate build in our CI, so we can profile easier
 - I also noticed from the profiling that the JSONEditor is effecting performance in a big way (more investigation required)