@pathmapper 2019-10-14
Thanks for the detailed explanation @orangemug . With the `max-zoom` test I'm seeing also a performance improvement. :tada: 

> 20ms is better than nothing

 18% increased performance sounds better :-)

> the JSONEditor is effecting performance in a big way (more investigation required)

Perhaps this annoying bug is related: https://github.com/maputnik/editor/issues/502