@pathmapper 2019-10-14
> I think I'll add a react profiling-in-production as a separate build in our CI, so we can profile easier

The intention is that we could use the react chrome extension on this CI build?