@orangemug 2019-10-15
> The intention is that we could use the react chrome extension on this CI build?

Yeah that's right, so what I'm thinking is that we'll create 2 artifacts from the CI builds one will be the normal build we use to show in demos currently. The other for performance testing we'll include the [react profiling-in-production](https://gist.github.com/bvaughn/25e6233aeb1b4f0cdb8d8366e54a3977#profiling-in-production), also add a debug option to maputnik were we can disable the renderer (mapbox-gl, ol) then we can get good metrics about the performance of our UI in isolation. 

> 18% increased performance sounds better :-)

👍 

> Perhaps this annoying bug is related: #502

I'll dig into the JSONEditor performance in another ticket.

Thanks again for all the help with this PR @pathmapper, having to justify my metrics really helped. I think we have a better approach moving forward.

Merging!